shopcode/requirements.txt
Django==1.6.5
Sphinx==1.2.2
Unipath==1.0
django-bootstrap3==4.11.0
django-tenant-schemas==1.4.8
